None Qubinode update and fixes for issues
When user choose not to setup system as qubinode. The installer will
attempt to discover the existing libvirt storage and network. If multiple
is found, the user will be prompted to pick one. For storage, the capacity
of the libvirt volume found will be check to make sure it meets our requirements.

Issues fixed, fixes Qubinode/qubinode-installers#83, fixes Qubinode/qubinode-installers#80.