Update qubinode_ansible.sh
added git to the installed packages. when you run ./qubinode-installer -m ansible, you get an error since it cannot download the required roles due to not having git installed. Same error occurs when you run the ./qubinode-installer default. I was not sure where to put this, so i went with the first possible option.